+++ list_client.c
@@ -59,7 +59,7 @@ int close_connection_list(struct temp_buffer temp_buff, struct shm_sel_repeat *s
     }
 }
 //è stata ricevuta tutta la lista aspetta il fin dal server per chiudere la trasmissione
-int wait_for_fin_list(struct temp_buffer temp_buff,struct shm_sel_repeat *shm) {
+long wait_for_fin_list(struct temp_buffer temp_buff,struct shm_sel_repeat *shm) {
     alarm(TIMEOUT);//chiusura temporizzata
     errno = 0;
     while (1) {
@@ -106,7 +106,7 @@ int wait_for_fin_list(struct temp_buffer temp_buff,struct shm_sel_repeat *shm) {
     }
 }
 //ricevuta dimensione della lista,manda messaggio di start per far capire al sender che è pronto a ricevere i dati
-int rcv_list(struct temp_buffer temp_buff,struct shm_sel_repeat *shm) {
+long rcv_list(struct temp_buffer temp_buff,struct shm_sel_repeat *shm) {
     alarm(TIMEOUT);
     send_message_in_window(temp_buff,shm, START,"START" );
     errno = 0;
@@ -164,7 +164,7 @@ int rcv_list(struct temp_buffer temp_buff,struct shm_sel_repeat *shm) {
     }
 }
 //manda messaggio get e aspetta messaggio contentente la dimensione della lista
-int wait_for_list_dimension(struct temp_buffer temp_buff,struct shm_sel_repeat *shm) {
+long wait_for_list_dimension(struct temp_buffer temp_buff,struct shm_sel_repeat *shm) {
     errno = 0;
     char *first;
     better_strcpy(temp_buff.payload, "list");
